@@ -18,35 +18,35 @@
     "cheerio": "^1.0.0",
     "cors": "^2.8.5",
     "express": "^4.21.2",
-    "multer": "^1.4.5-lts.1",
-    "pg": "^8.13.3",
+    "multer": "^1.4.5-lts.2",
+    "pg": "^8.14.1",
     "pg-hstore": "^2.3.4",
-    "sequelize": "^6.37.5",
+    "sequelize": "^6.37.6",
     "sequelize-cli": "^6.6.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@babel/preset-env": "^7.26.0",
-    "@babel/preset-typescript": "^7.26.0",
+    "@babel/core": "^7.26.10",
+    "@babel/preset-env": "^7.26.9",
+    "@babel/preset-typescript": "^7.27.0",
     "@eslint/compat": "^1.2.7",
-    "@eslint/eslintrc": "^3.3.0",
-    "@eslint/js": "^9.19.0",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.23.0",
     "@types/body-parser": "^1.19.5",
     "@types/cors": "^2.8.17",
-    "@types/express": "^5.0.0",
+    "@types/express": "^5.0.1",
     "@types/multer": "^1.4.12",
-    "@types/node": "^22.10.5",
+    "@types/node": "^22.13.13",
     "@types/pg": "^8.11.11",
     "@types/sequelize": "^4.28.20",
     "@types/webpack": "^5.28.5",
-    "@typescript-eslint/parser": "^8.22.0",
+    "@typescript-eslint/parser": "^8.28.0",
     "babel-loader": "^10.0.0",
     "npm-run-all": "^4.1.5",
-    "ts-loader": "^9.5.1",
+    "ts-loader": "^9.5.2",
     "ts-node": "^10.9.2",
     "tsconfig-paths-webpack-plugin": "^4.2.0",
-    "typescript": "^5.7.3",
-    "wait-on": "^8.0.2",
+    "typescript": "^5.8.2",
+    "wait-on": "^8.0.3",
     "webpack": "^5.98.0",
     "webpack-cli": "^6.0.1",
     "worker-plugin": "^5.0.1"
