@@ -1,21 +1,27 @@
-/** @type {import('next').NextConfig} */
+/**
+ * @type {import('next').NextConfig}
+ */
 const nextConfig = {
   images: {
     remotePatterns: [
       {
+        protocol: 'https',
         hostname: 'tecdn.b-cdn.net',
       },
       {
+        protocol: 'https',
         hostname: 'flagcdn.com',
       },
       {
+        protocol: 'https',
         hostname: 'upload.wikimedia.org',
       },
       {
+        protocol: 'https',
         hostname: 'flagpedia.net',
       },
     ],
   },
 }
 
-module.exports = nextConfig
+export default nextConfig
