@@ -2,7 +2,7 @@ import React, { useEffect, useState, useMemo } from 'react'
 import Image, { ImageProps } from 'next/image'
 
 type ImageAutoSizeProps = Omit<ImageProps, 'src'> & {
-  imageBlob?: Blob;
+  imageBlob?: Blob | null;
   fallbackSrc?: string;
   src?: string;
 }
@@ -12,7 +12,7 @@ const ImageAutoSize: React.FC<ImageAutoSizeProps> = (props) => {
   const [objectUrl, setObjectUrl] = useState<string | null>(null)
 
   useEffect(() => {
-    if (imageBlob) {
+    if (imageBlob && imageBlob instanceof Blob) {
       const url = URL.createObjectURL(imageBlob)
       setObjectUrl(url)
 
@@ -25,6 +25,7 @@ const ImageAutoSize: React.FC<ImageAutoSizeProps> = (props) => {
     }
   }, [imageBlob])
 
+  // Memoize the image source to prevent unnecessary re-renders
   const imageSrc = useMemo(() => objectUrl || src || fallbackSrc || '', [objectUrl, src, fallbackSrc])
 
   return (
@@ -35,8 +36,10 @@ const ImageAutoSize: React.FC<ImageAutoSizeProps> = (props) => {
       width={width}
       height={height}
       style={{ maxWidth: width, maxHeight: height, width: width, height: height, ...style }}
+      // Remove lazy loading and priority attributes as they may be causing issues
     />
   )
 }
 
+// Use React.memo to prevent unnecessary re-renders of this component
 export default React.memo(ImageAutoSize)
