@@ -38,11 +38,19 @@ const Pagination: React.FC<PaginationProps> = ({ totalItems, onPageChange, limit
 
   return (
     <div className="flex justify-between mt-4">
-      <button onClick={handlePreviousPage} disabled={currentPage === 1} className="bg-gray-500 text-white mr-4 px-4 py-2 rounded hover:bg-gray-700 disabled:bg-gray-300">
+      <button 
+        onClick={handlePreviousPage} 
+        disabled={currentPage === 1} 
+        className="bg-gray-500 text-white mr-4 px-4 py-2 rounded hover:bg-gray-700 disabled:bg-gray-300"
+      >
         Previous
       </button>
       <span className="text-gray-700 py-2">Page {currentPage} of {totalPages}</span>
-      <button onClick={handleNextPage} disabled={currentPage >= totalPages} className="bg-gray-500 text-white px-4 py-2 ml-4 rounded hover:bg-gray-700 disabled:bg-gray-300">
+      <button 
+        onClick={handleNextPage} 
+        disabled={currentPage >= totalPages} 
+        className="bg-gray-500 text-white px-4 py-2 ml-4 rounded hover:bg-gray-700 disabled:bg-gray-300"
+      >
         Next
       </button>
     </div>
