@@ -1,7 +1,7 @@
 "use client"
 
 import { useRouter } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import { useEffect, useState, useCallback } from 'react'
 import { fetchCountries } from '../api/CountryApi'
 import { fetchPlayers, deletePlayer } from '../api/PlayersApi'
 import { getAttributeBgColor, getRoleBgColor, Player } from '../api/models/Player'
@@ -15,8 +15,9 @@ import { ItemsWithPagination } from '../api/models/types'
 import SectionHeader from '../base/SectionHeader'
 import ImageAutoSize from '../base/ImageAutoSize'
 
+const DEFAULT_LIMIT_VALUE_PLAYER_LIST = 5 // Return to original value
+
 export default function ListPlayers() {
-  const LIMIT_VALUE_PLAYER_LIST = 5
   const router = useRouter()
   const [players, setPlayers] = useState<Player[]>([])
   const [playerToEdit, setPlayerToEdit] = useState<Player | null>(null)
@@ -25,6 +26,35 @@ export default function ListPlayers() {
   const [isEditActionOpened, setIsEditActionOpened] = useState<boolean>(false)
   const [countriesToFlagMap, setCountriesToFlagMap] = useState<Record<string, string>>({})
   const [totalItems, setTotalItems] = useState(0)
+  const [isLoading, setIsLoading] = useState(true)
+  const [limitValue, setLimitValue] = useState(DEFAULT_LIMIT_VALUE_PLAYER_LIST)
+
+  // Simplify the refreshListData function
+  const refreshListData = useCallback(async (data: ItemsWithPagination<Player>) => {
+    setIsLoading(true)
+    
+    try {
+      const playerToTeam: Record<string, Team> = {}
+      
+      const teamFetchPromises = data.items.map((player) =>
+        fetchTeam(player.team_id, team => {
+          if (player.id) {
+            playerToTeam[player.id] = team
+          }
+        }),
+      )
+  
+      await Promise.all(teamFetchPromises)
+      
+      setPlayerToTeam(playerToTeam)
+      setTotalItems(data.total)
+      setPlayers(data.items)
+    } catch (error) {
+      console.error("Error refreshing player data:", error)
+    } finally {
+      setIsLoading(false)
+    }
+  }, [])
 
   useEffect(() => {
     fetchCountries((countries) => {
@@ -34,73 +64,62 @@ export default function ListPlayers() {
       })
       setCountriesToFlagMap(countriesToFlagMap)
     })
+    
+    fetchPlayers(refreshListData, limitValue)
+  }, [refreshListData, limitValue])
 
-    fetchPlayers(refreshListData, LIMIT_VALUE_PLAYER_LIST) 
-  }, [])
-
-  const refreshListData = async (data: ItemsWithPagination<Player>) => {
-    const playerToTeam: Record<number, Team> = {}
-
-    const teamFetchPromises = data.items.map((player) =>
-      fetchTeam(player.team_id, team => {
-        if (player.id) {
-          playerToTeam[player.id] = team
-        }
-      }),
-    )
-
-    await Promise.all(teamFetchPromises)
-
-    setPlayerToTeam(playerToTeam)
-    setTotalItems(data.total)
-    setPlayers(data.items)
-  }
-
-  const openNewPlayerModal = (e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
+  const openNewPlayerModal = useCallback((e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
     e.preventDefault()
     setActionPlayerModalOpened(true)
-  }
+  }, [])
 
-  const closeActionPlayerModal = () => {
+  const closeActionPlayerModal = useCallback(() => {
     setIsEditActionOpened(false)
     setPlayerToEdit(null)
     setActionPlayerModalOpened(false)
-    fetchPlayers(refreshListData, LIMIT_VALUE_PLAYER_LIST)
-  }
+    fetchPlayers(refreshListData, limitValue)
+  }, [refreshListData, limitValue])
 
-  const handleView = (player: Player) => {
-    // Send user to player details page
+  const handleView = useCallback((player: Player) => {
     router.push(`/players/${player.id}`)
-  }
+  }, [router])
 
-  const handleEdit = (player: Player) => {
-    // Use same modal as NewPlayerModal but with prefilled data
+  const handleEdit = useCallback((player: Player) => {
     setPlayerToEdit(player)
     setIsEditActionOpened(true)
     setActionPlayerModalOpened(true)
-  }
+  }, [])
 
-  const handleDelete = (player: Player) => {
-    // Show confirm dialog and if confirmed delete player
+  const handleDelete = useCallback((player: Player) => {
     const confirmed = confirm(`Are you sure you want to delete player '${player.nickname}'?`)
     if(!confirmed) return
 
     deletePlayer(player, () => {
-      fetchPlayers(refreshListData)
+      fetchPlayers(refreshListData, limitValue)
     })
-    
-  }
+  }, [refreshListData, limitValue])
 
-  const handlePageChange = (limit: number, offset: number) => {
+  const handlePageChange = useCallback((limit: number, offset: number) => {
+    setLimitValue(limit)
+    setIsLoading(true)
     fetchPlayers(refreshListData, limit, offset)
-  }
+  }, [refreshListData])
 
-  // List all players in a table/grid
   return (
     <>
-      <PlayerActionModal isOpen={actionPlayerModalOpened} isEdit={isEditActionOpened} object={playerToEdit} onClose={closeActionPlayerModal} />
+      <PlayerActionModal 
+        isOpen={actionPlayerModalOpened} 
+        isEdit={isEditActionOpened} 
+        object={playerToEdit} 
+        onClose={closeActionPlayerModal} 
+      />
       <div className="flex min-h-screen flex-col items-center p-24">
-        <SectionHeader title="Players" action={openNewPlayerModal} actionText="New Player" />
+        <SectionHeader 
+          title="Players" 
+          action={openNewPlayerModal} 
+          actionText="New Player" 
+        />
+        
         <table className="min-w-full divide-y divide-gray-200">
           <thead>
             <tr>
@@ -134,63 +153,86 @@ export default function ListPlayers() {
             </tr>
           </thead>
           <tbody className="bg-white divide-y divide-gray-200">
-            {players.map((player) => (
-              <tr key={player.id}>
-                <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.id}</td>
-                <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.nickname}</td>
-                <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.full_name}</td>
-                <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.age}</td>
-                <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                  {player.country && (
-                    <span className="flex items-center">
-                      <ImageAutoSize src={countriesToFlagMap[player.country]} alt={player.country} width={32} height={16} className="mr-2" />
-                      {player.country}
-                    </span>
-                  )}
-                </td>
-                <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                  {playerToTeam && playerToTeam[String(player.id)] ? (
-                    <span className="flex items-center">
-                      <ImageAutoSize 
-                        imageBlob={playerToTeam[String(player.id)].logo_image_file as Blob}
-                        fallbackSrc="/images/nologo.svg"
-                        alt={playerToTeam[String(player.id)].short_name} 
-                        width={32} 
-                        height={32} 
-                        className="mr-2" />
-                      {playerToTeam[String(player.id)].short_name}
-                    </span>
-                  ) : (
-                    'No Team'
-                  )}
-                </td>
-                <td>
-                  <span className={getRoleBgColor(player.role)}>
-                    {player.role}
-                  </span>
-                </td>
-                <td className="px-6 py-4 whitespace-normal text-sm font-medium text-gray-500">
-                  <div className="grid grid-cols-1 sm:grid-cols-1 md:grid-cols-4 gap-1">
-                    {Object.entries(player.player_attributes).map(([key, value]) => {
-                      return (
-                        <div key={key} className="flex items-center space-x-1">
-                          <span className={`w-6 h-6 flex items-center justify-center rounded text-xs text-white ${getAttributeBgColor(value)}`}>{value}</span>
-                          <span className="text-xs text-gray-900 truncate">{asWord(key)}</span>
-                        </div>
-                      )
-                    })}
-                  </div>
-                </td>
-                <td className="py-4 whitespace-nowrap text-sm text-left text-gray-900 w-auto">
-                  <button onClick={() => handleView(player)} className="text-blue-600 hover:text-blue-900 p0">👀</button>
-                  <button onClick={() => handleEdit(player)} className="text-blue-600 hover:text-blue-900 p0">✏️</button>
-                  <button onClick={() => handleDelete(player)} className="text-red-600 hover:text-red-900 p0">🗑️</button>
+            {isLoading ? (
+              // Display placeholder loading rows
+              Array.from({ length: 5 }).map((_, index) => (
+                <tr key={`loading-${index}`}>
+                  <td colSpan={9}>
+                    <div className="animate-pulse h-16 bg-gray-100 my-1"></div>
+                  </td>
+                </tr>
+              ))
+            ) : players.length === 0 ? (
+              <tr>
+                <td colSpan={9} className="px-3 py-4 text-center text-gray-500">
+                  No players found
                 </td>
               </tr>
-            ))}
+            ) : (
+              // Inline the player rows instead of using a separate component
+              players.map((player) => (
+                <tr key={player.id}>
+                  <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.id}</td>
+                  <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.nickname}</td>
+                  <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.full_name}</td>
+                  <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{player.age}</td>
+                  <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
+                    {player.country && (
+                      <span className="flex items-center">
+                        <ImageAutoSize src={countriesToFlagMap[player.country]} alt={player.country} width={32} height={16} className="mr-2" />
+                        {player.country}
+                      </span>
+                    )}
+                  </td>
+                  <td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
+                    {playerToTeam && playerToTeam[String(player.id)] ? (
+                      <span className="flex items-center">
+                        <ImageAutoSize 
+                          imageBlob={playerToTeam[String(player.id)].logo_image_file as Blob}
+                          fallbackSrc="/images/nologo.svg"
+                          alt={playerToTeam[String(player.id)].short_name} 
+                          width={32} 
+                          height={32} 
+                          className="mr-2" />
+                        {playerToTeam[String(player.id)].short_name}
+                      </span>
+                    ) : (
+                      'No Team'
+                    )}
+                  </td>
+                  <td>
+                    <span className={getRoleBgColor(player.role)}>
+                      {player.role}
+                    </span>
+                  </td>
+                  <td className="px-6 py-4 whitespace-normal text-sm font-medium text-gray-500">
+                    <div className="grid grid-cols-1 sm:grid-cols-1 md:grid-cols-4 gap-1">
+                      {Object.entries(player.player_attributes).map(([key, value]) => {
+                        return (
+                          <div key={key} className="flex items-center space-x-1">
+                            <span className={`w-6 h-6 flex items-center justify-center rounded text-xs text-white ${getAttributeBgColor(value)}`}>{value}</span>
+                            <span className="text-xs text-gray-900 truncate">{asWord(key)}</span>
+                          </div>
+                        )
+                      })}
+                    </div>
+                  </td>
+                  <td className="py-4 whitespace-nowrap text-sm text-left text-gray-900 w-auto">
+                    <button onClick={() => handleView(player)} className="text-blue-600 hover:text-blue-900 p-0 mr-1">👀</button>
+                    <button onClick={() => handleEdit(player)} className="text-blue-600 hover:text-blue-900 p-0 mr-1">✏️</button>
+                    <button onClick={() => handleDelete(player)} className="text-red-600 hover:text-red-900 p-0">🗑️</button>
+                  </td>
+                </tr>
+              ))
+            )}
           </tbody>
         </table>
-        <Pagination totalItems={totalItems} onPageChange={handlePageChange} limitValue={LIMIT_VALUE_PLAYER_LIST} />
+        
+        <Pagination 
+          totalItems={totalItems} 
+          onPageChange={handlePageChange} 
+          limitValue={limitValue} 
+        />
       </div>
     </>
   )
